Move some declarations to gresource.h
